amplify/backend/function/fetchCryptoPrices/src/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
boto3==1.28.0
botocore==1.31.0 